refactor: replace bespoke status color maps with Badge's color API
System job and content pack status badges each had their own
styled(Badge) wrapper computing colors from theme.colors.variant.dark
by hand. Use color/variant/dot instead, so they pick up the same
palette as every other status badge. 'info' has no equivalent in the
new color set, so queued/updatable map to gray.
